requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
supabase>=2.0.3
pyjwt>=2.8.0
bcrypt>=4.1.1
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
requests>=2.31.0


